wall climbing added?
idk what's happened, idk how it's fixed, but it is. i will not question it, nor will i go back and try and fix it. as long as it works my sanity works too

(i still need to disable wall climbing for a short period of time which is why im commiting)